Create Landsat 7 NDWI and Scan Line Fix - GEE
This code does not mask for clouds, but it does its best to substitute in Landsat 5 data where fringes of Landsat 7 data occur. It's not foolproof, but it's better than having fringes. This code also specifically draws from the Landsat 7 NDWI calculation, but it can be applied to other Landsat 7 datasets. Just be sure to change dataset variables and bands.